[🧹 REFACTOR] 메인 페이지 경로 변경 '/'에서 로그인 페이지로 리디렉션
CLOSE #96
