+++ handtris/src/app/page.tsx
@@ -1,30 +1,14 @@
 "use client";
-import { useEffect, useRef, useState } from "react";
-import { motion, AnimatePresence } from "framer-motion";
-import Image from "next/image";
-import { useRouter } from "next/navigation";
-import { menuClickSound, menuHoverSound } from "@/hook/howl";
-import { useMusic } from "@/components/MusicProvider";
-import SplashScreen from "@/components/SplashScreen";
 
-export default function HomePage() {
-  const [selected, setSelected] = useState<string | null>(null);
+import React, { useState, useEffect } from "react";
+import LoginForm from "@/components/LoginForm";
+import SplashScreen from "@/components/SplashScreen";
+import { useMusic } from "@/components/MusicProvider";
+
+const LoginPage = () => {
   const [showSplash, setShowSplash] = useState(true);
-  const router = useRouter();
-  const mainRef = useRef<HTMLElement>(null);
   const { isMusicPlaying, toggleMusic } = useMusic();
 
-  const handleSelect = (option: string, path: string) => {
-    setSelected(option);
-    menuClickSound();
-    setTimeout(() => {
-      router.push(path);
-    }, 1000);
-  };
-
-  const handleHover = () => {
-    menuHoverSound();
-  };
   useEffect(() => {
     const hasSeenSplash = localStorage.getItem("hasSeenSplash");
     if (hasSeenSplash) {
@@ -42,105 +26,16 @@ export default function HomePage() {
   }
 
   return (
-    <main
-      tabIndex={-1}
-      ref={mainRef}
-      className="flex min-h-screen flex-col items-center justify-center p-8"
-    >
+    <div className="flex h-screen flex-col items-center justify-center p-4">
       <button
         onClick={toggleMusic}
         className="absolute top-4 right-4 bg-white bg-opacity-20 p-2 rounded-full"
       >
         {isMusicPlaying ? "🔇" : "🔊"}
       </button>
-      <div className="flex w-full max-w-2xl flex-col space-y-4 rounded-xl border border-white border-opacity-20 bg-white bg-opacity-10 p-8 shadow-2xl backdrop-blur-lg backdrop-filter">
-        <AnimatePresence>
-          {selected === null && (
-            <>
-              <motion.div
-                className="w-full"
-                initial={{ opacity: 0, x: -100 }}
-                animate={{ opacity: 1, x: 0 }}
-                exit={{ opacity: 0, x: 1000 }}
-                transition={{ duration: 0.5, delay: 0.1 }}
-              >
-                <button
-                  className="w-full transform rounded-lg border border-pink-500 border-opacity-40 bg-white bg-opacity-30 px-8 py-4 font-bold text-pink-500 shadow-lg backdrop-blur-sm backdrop-filter transition hover:scale-105"
-                  onClick={() => handleSelect("lobby", "/lobby")}
-                  onMouseEnter={handleHover}
-                >
-                  <div className="flex items-center justify-between">
-                    <Image
-                      src="/image/start.png"
-                      alt="option"
-                      width={150}
-                      height={150}
-                    />
-                    <div className="flex flex-col items-start">
-                      <span className="text-left">GAME PLAY</span>
-                      <p className="text-sm">
-                        Play online with friends and foes
-                      </p>
-                    </div>
-                  </div>
-                </button>
-              </motion.div>
-              <motion.div
-                className="w-full"
-                initial={{ opacity: 0, x: -100 }}
-                animate={{ opacity: 1, x: 0 }}
-                exit={{ opacity: 0, x: 1000 }}
-                transition={{ duration: 0.5, delay: 0.2 }}
-              >
-                <button
-                  className="w-full transform rounded-lg border border-green-500 border-opacity-40 bg-white bg-opacity-30 px-8 py-4 font-bold text-green-500 shadow-lg backdrop-blur-sm backdrop-filter transition hover:scale-105"
-                  onClick={() => handleSelect("config", "/config")}
-                  onMouseEnter={handleHover}
-                >
-                  <div className="flex items-center justify-between">
-                    <Image
-                      src="/image/config.png"
-                      alt="option"
-                      width={150}
-                      height={150}
-                    />
-                    <div className="flex flex-col items-start">
-                      <span className="text-left">CONFIG</span>
-                      <p className="text-sm">Tweak your settings</p>
-                    </div>
-                  </div>
-                </button>
-              </motion.div>
-              <motion.div
-                className="w-full"
-                initial={{ opacity: 0, x: -100 }}
-                animate={{ opacity: 1, x: 0 }}
-                exit={{ opacity: 0, x: 1000 }}
-                transition={{ duration: 0.5, delay: 0.3 }}
-              >
-                <button
-                  className="w-full transform rounded-lg border border-blue-500 border-opacity-40 bg-white bg-opacity-30 px-8 py-4 font-bold text-blue-500 shadow-lg backdrop-blur-sm backdrop-filter transition hover:scale-105"
-                  onClick={() => handleSelect("about", "/play/tetris")}
-                  onMouseEnter={handleHover}
-                >
-                  <div className="flex items-center justify-between">
-                    <Image
-                      src="/image/tools.png"
-                      alt="option"
-                      width={150}
-                      height={150}
-                    />
-                    <div className="flex flex-col items-start">
-                      <span className="text-left">ABOUT</span>
-                      <p className="text-sm">Learn more about us</p>
-                    </div>
-                  </div>
-                </button>
-              </motion.div>
-            </>
-          )}
-        </AnimatePresence>
-      </div>
-    </main>
+      <LoginForm />
+    </div>
   );
-}
+};
+
+export default LoginPage;
